refactor session_repository (tested)
