+++ todo_app/app.py
@@ -1,25 +1,20 @@
-from todo_app.trello_wrapper import fake_trello_wrapper, real_trello_wrapper
+from todo_app.data.trello_wrapper import fake_trello_wrapper, real_trello_wrapper
 from flask import Flask, render_template, request, redirect
-from todo_app.data.session_items import delete_item, get_items, save_item, add_item, get_item
+from todo_app.data.session_repository import session_repository
 
 from todo_app.flask_config import Config
 
 app = Flask(__name__)
 app.config.from_object(Config)
+repository = session_repository()
 
 
 @app.route('/')
 def index():
     print("index()")
-    items = sorted(get_items(), key=lambda item: item["status"], reverse=True)
-    for i in get_items():
-        print(i)
-
-    trello = fake_trello_wrapper()
-    print(trello.getAllBoards())
-    print(trello.getAllCards())
-    print(trello.getAllLists())
-
+    items = sorted(repository.get_items(), key=lambda item: item["status"], reverse=True)
+    for i in items:
+        print(i)        
     return render_template('index.html', items=items) 
 
 @app.route('/items', methods=['POST'])
@@ -27,22 +22,22 @@ def add_new_item():
     print("add_new_item")
     title = request.form['title']
     print(title)
-    add_item(title)
+    repository.add_item(title)
     return redirect('/')
 
 @app.route('/mark_as_done/<itemId>', methods=['POST']) 
 def mark_as_done(itemId): 
     print(f"mark_as_done: itemId = {itemId}")
-    item = get_item(itemId)
+    item = repository.get_item(itemId)
     item["status"] = "Done"
-    save_item(item)
+    repository.save_item(item)
     return redirect('/')
 
 @app.route('/delete/<itemId>', methods=['POST']) 
 def delete(itemId): 
     try:
         print(f"delete: itemId = {itemId}")
-        delete_item(itemId)        
+        repository.delete_item(itemId)        
     except Exception as error:
         print(f"Exception: {error}")
     finally:
